training_patterns/requirements/requirements_train.txt
dvc>=1.4.0
future>=0.18.2
importlib-metadata>=1.7.0
joblib>=0.15.1
numpy>=1.18.5
onnx>=1.7.0
onnxruntime>=1.4.0
PyYAML>=5.3.1
scikit-learn==0.23.1
skl2onnx>=1.7.0
